no need to cast va_arg output

+++ echo.c
@@ -654,7 +654,7 @@ eformat(fp, ap)
 				c = *fp++;
 				switch (c) {
 				case 'd':
-					eputl((long)va_arg(ap, long), 10);
+					eputl(va_arg(ap, long), 10);
 					break;
 				default:
 					eputc(c);
